fix: improve navigation components and fix button functionality - Create dedicated Navigation components for better maintainability - Separate DesktopNavigation and MobileNavigation components - Ensure all navigation links work properly - Fix potential JavaScript issues preventing button clicks - All builds pass successfully
